+++ packages/tailwindcss/src/utilities.test.ts
@@ -145,7 +145,7 @@ test('inset', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadow-sm: inset 0 1px 1px #0000000d;
       --inset-shadowned: 1940px;
@@ -295,7 +295,7 @@ test('inset-x', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -382,7 +382,7 @@ test('inset-y', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -469,7 +469,7 @@ test('start', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -556,7 +556,7 @@ test('end', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -644,7 +644,7 @@ test('top', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -731,7 +731,7 @@ test('right', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -818,7 +818,7 @@ test('bottom', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -905,7 +905,7 @@ test('left', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --inset-shadowned: 1940px;
     }
@@ -1444,7 +1444,7 @@ test('margin', async () => {
       ['m-auto', 'm-4', 'm-[4px]', '-m-4', '-m-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -1498,7 +1498,7 @@ test('mx', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -1584,7 +1584,7 @@ test('my', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -1670,7 +1670,7 @@ test('mt', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -1756,7 +1756,7 @@ test('ms', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -1842,7 +1842,7 @@ test('me', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -1928,7 +1928,7 @@ test('mr', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -2014,7 +2014,7 @@ test('mb', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -2100,7 +2100,7 @@ test('ml', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -2173,7 +2173,7 @@ test('margin sort order', async () => {
       ['mb-4', 'me-4', 'mx-4', 'ml-4', 'ms-4', 'm-4', 'mr-4', 'mt-4', 'my-4'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -2477,7 +2477,7 @@ test('aspect-ratio', async () => {
       ['aspect-video', 'aspect-[10/9]', 'aspect-4/3'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --aspect-video: 16 / 9;
     }
 
@@ -2531,7 +2531,7 @@ test('size', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -2624,7 +2624,7 @@ test('width', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --width-xl: 36rem;
     }
@@ -2730,7 +2730,7 @@ test('min-width', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --container-xl: 36rem;
     }
@@ -2806,7 +2806,7 @@ test('max-width', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
       --container-xl: 36rem;
     }
@@ -2881,7 +2881,7 @@ test('height', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -2983,7 +2983,7 @@ test('min-height', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -3075,7 +3075,7 @@ test('max-height', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -3160,7 +3160,7 @@ describe('container', () => {
         ['w-1/2', 'container', 'max-w-[var(--breakpoint-sm)]'],
       ),
     ).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --breakpoint-sm: 40rem;
         --breakpoint-md: 48rem;
         --breakpoint-lg: 64rem;
@@ -3303,7 +3303,7 @@ describe('container', () => {
         ['w-1/2', 'container', 'max-w-[var(--breakpoint-sm)]'],
       ),
     ).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --breakpoint-sm: 40rem;
         --breakpoint-md: 48rem;
         --breakpoint-lg: 64rem;
@@ -3500,7 +3500,7 @@ test('flex-basis', async () => {
       ['basis-auto', 'basis-full', 'basis-xl', 'basis-11/12', 'basis-[123px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --container-xl: 36rem;
     }
 
@@ -3604,7 +3604,7 @@ test('border-spacing', async () => {
       ['border-spacing-1', 'border-spacing-[123px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-1: .25rem;
     }
 
@@ -3655,7 +3655,7 @@ test('border-spacing-x', async () => {
       ['border-spacing-x-1', 'border-spacing-x-[123px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-1: .25rem;
     }
 
@@ -3704,7 +3704,7 @@ test('border-spacing-y', async () => {
       ['border-spacing-y-1', 'border-spacing-y-[123px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-1: .25rem;
     }
 
@@ -4045,7 +4045,7 @@ test('translate-x', async () => {
       ['translate-x-full', '-translate-x-full', 'translate-x-px', '-translate-x-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
     }
 
@@ -4173,7 +4173,7 @@ test('translate-y', async () => {
       ['translate-y-full', '-translate-y-full', 'translate-y-px', '-translate-y-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
     }
 
@@ -5113,7 +5113,7 @@ test('perspective', async () => {
       ['perspective-normal', 'perspective-dramatic', 'perspective-none', 'perspective-[456px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --perspective-dramatic: 100px;
       --perspective-normal: 500px;
     }
@@ -5199,7 +5199,7 @@ test('cursor', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --cursor-custom: url("/my-cursor.png");
     }
 
@@ -5751,7 +5751,7 @@ test('scroll-m', async () => {
       ['scroll-m-4', 'scroll-m-[4px]', '-scroll-m-4', '-scroll-m-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -5794,7 +5794,7 @@ test('scroll-mx', async () => {
       ['scroll-mx-4', 'scroll-mx-[4px]', '-scroll-mx-4', '-scroll-mx-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -5837,7 +5837,7 @@ test('scroll-my', async () => {
       ['scroll-my-4', 'scroll-my-[4px]', '-scroll-my-4', '-scroll-my-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -5880,7 +5880,7 @@ test('scroll-ms', async () => {
       ['scroll-ms-4', 'scroll-ms-[4px]', '-scroll-ms-4', '-scroll-ms-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -5923,7 +5923,7 @@ test('scroll-me', async () => {
       ['scroll-me-4', 'scroll-me-[4px]', '-scroll-me-4', '-scroll-me-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -5966,7 +5966,7 @@ test('scroll-mt', async () => {
       ['scroll-mt-4', 'scroll-mt-[4px]', '-scroll-mt-4', '-scroll-mt-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6009,7 +6009,7 @@ test('scroll-mr', async () => {
       ['scroll-mr-4', 'scroll-mr-[4px]', '-scroll-mr-4', '-scroll-mr-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6052,7 +6052,7 @@ test('scroll-mb', async () => {
       ['scroll-mb-4', 'scroll-mb-[4px]', '-scroll-mb-4', '-scroll-mb-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6095,7 +6095,7 @@ test('scroll-ml', async () => {
       ['scroll-ml-4', 'scroll-ml-[4px]', '-scroll-ml-4', '-scroll-ml-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6138,7 +6138,7 @@ test('scroll-p', async () => {
       ['scroll-p-4', 'scroll-p-[4px]', '-scroll-p-4', '-scroll-p-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6173,7 +6173,7 @@ test('scroll-px', async () => {
       ['scroll-px-4', 'scroll-px-[4px]', '-scroll-px-4', '-scroll-px-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6208,7 +6208,7 @@ test('scroll-py', async () => {
       ['scroll-py-4', 'scroll-py-[4px]', '-scroll-py-4', '-scroll-py-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6243,7 +6243,7 @@ test('scroll-ps', async () => {
       ['scroll-ps-4', 'scroll-ps-[4px]', '-scroll-ps-4', '-scroll-ps-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6278,7 +6278,7 @@ test('scroll-pe', async () => {
       ['scroll-pe-4', 'scroll-pe-[4px]', '-scroll-pe-4', '-scroll-pe-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6313,7 +6313,7 @@ test('scroll-pt', async () => {
       ['scroll-pt-4', 'scroll-pt-[4px]', '-scroll-pt-4', '-scroll-pt-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6348,7 +6348,7 @@ test('scroll-pr', async () => {
       ['scroll-pr-4', 'scroll-pr-[4px]', '-scroll-pr-4', '-scroll-pr-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6383,7 +6383,7 @@ test('scroll-pb', async () => {
       ['scroll-pb-4', 'scroll-pb-[4px]', '-scroll-pb-4', '-scroll-pb-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6418,7 +6418,7 @@ test('scroll-pl', async () => {
       ['scroll-pl-4', 'scroll-pl-[4px]', '-scroll-pl-4', '-scroll-pl-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -6599,7 +6599,7 @@ test('columns', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --container-3xs: 16rem;
       --container-7xl: 80rem;
     }
@@ -7467,7 +7467,7 @@ test('gap', async () => {
       ['gap-4', 'gap-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -7494,7 +7494,7 @@ test('gap-x', async () => {
       ['gap-x-4', 'gap-x-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -7523,7 +7523,7 @@ test('gap-y', async () => {
       ['gap-y-4', 'gap-y-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -7552,7 +7552,7 @@ test('space-x', async () => {
       ['space-x-4', 'space-x-[4px]', '-space-x-4'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -7595,7 +7595,7 @@ test('space-y', async () => {
       ['space-y-4', 'space-y-[4px]', '-space-y-4'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing-4: 1rem;
     }
 
@@ -7732,7 +7732,7 @@ test('divide-x with custom default border width', async () => {
       ['divide-x'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --default-border-width: 2px;
     }
 
@@ -7838,7 +7838,7 @@ test('divide-y with custom default border width', async () => {
       ['divide-y'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --default-border-width: 2px;
     }
 
@@ -7971,7 +7971,7 @@ test('accent', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -8086,7 +8086,7 @@ test('caret', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -8199,7 +8199,7 @@ test('divide-color', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -8863,7 +8863,7 @@ test('rounded', async () => {
       ['rounded', 'rounded-full', 'rounded-none', 'rounded-sm', 'rounded-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-sm: .125rem;
       --radius: .25rem;
     }
@@ -8919,7 +8919,7 @@ test('rounded-s', async () => {
       ['rounded-s', 'rounded-s-full', 'rounded-s-none', 'rounded-s-sm', 'rounded-s-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -8982,7 +8982,7 @@ test('rounded-e', async () => {
       ['rounded-e', 'rounded-e-full', 'rounded-e-none', 'rounded-e-sm', 'rounded-e-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9045,7 +9045,7 @@ test('rounded-t', async () => {
       ['rounded-t', 'rounded-t-full', 'rounded-t-none', 'rounded-t-sm', 'rounded-t-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9112,7 +9112,7 @@ test('rounded-r', async () => {
       ['rounded-r', 'rounded-r-full', 'rounded-r-none', 'rounded-r-sm', 'rounded-r-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9179,7 +9179,7 @@ test('rounded-b', async () => {
       ['rounded-b', 'rounded-b-full', 'rounded-b-none', 'rounded-b-sm', 'rounded-b-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9246,7 +9246,7 @@ test('rounded-l', async () => {
       ['rounded-l', 'rounded-l-full', 'rounded-l-none', 'rounded-l-sm', 'rounded-l-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9313,7 +9313,7 @@ test('rounded-ss', async () => {
       ['rounded-ss', 'rounded-ss-full', 'rounded-ss-none', 'rounded-ss-sm', 'rounded-ss-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9371,7 +9371,7 @@ test('rounded-se', async () => {
       ['rounded-se', 'rounded-se-full', 'rounded-se-none', 'rounded-se-sm', 'rounded-se-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9429,7 +9429,7 @@ test('rounded-ee', async () => {
       ['rounded-ee', 'rounded-ee-full', 'rounded-ee-none', 'rounded-ee-sm', 'rounded-ee-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9487,7 +9487,7 @@ test('rounded-es', async () => {
       ['rounded-es', 'rounded-es-full', 'rounded-es-none', 'rounded-es-sm', 'rounded-es-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9545,7 +9545,7 @@ test('rounded-tl', async () => {
       ['rounded-tl', 'rounded-tl-full', 'rounded-tl-none', 'rounded-tl-sm', 'rounded-tl-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9605,7 +9605,7 @@ test('rounded-tr', async () => {
       ['rounded-tr', 'rounded-tr-full', 'rounded-tr-none', 'rounded-tr-sm', 'rounded-tr-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9665,7 +9665,7 @@ test('rounded-br', async () => {
       ['rounded-br', 'rounded-br-full', 'rounded-br-none', 'rounded-br-sm', 'rounded-br-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9725,7 +9725,7 @@ test('rounded-bl', async () => {
       ['rounded-bl', 'rounded-bl-full', 'rounded-bl-none', 'rounded-bl-sm', 'rounded-bl-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --radius-none: 0px;
       --radius-full: 9999px;
       --radius-sm: .125rem;
@@ -9928,7 +9928,7 @@ test('border with custom default border width', async () => {
       ['border'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --default-border-width: 2px;
     }
 
@@ -10089,7 +10089,7 @@ test('bg', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -10702,7 +10702,7 @@ test('from', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -10921,7 +10921,7 @@ test('via', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -11150,7 +11150,7 @@ test('to', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -11700,7 +11700,7 @@ test('fill', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -11821,7 +11821,7 @@ test('stroke', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -12074,7 +12074,7 @@ test('p', async () => {
       ['p-1', 'p-4', 'p-99', 'p-big', 'p-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12115,7 +12115,7 @@ test('px', async () => {
       ['px-1', 'px-99', 'px-2.5', 'px-big', 'px-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12156,7 +12156,7 @@ test('py', async () => {
       ['py-1', 'py-4', 'py-99', 'py-big', 'py-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12197,7 +12197,7 @@ test('pt', async () => {
       ['pt-1', 'pt-4', 'pt-99', 'pt-big', 'pt-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12238,7 +12238,7 @@ test('ps', async () => {
       ['ps-1', 'ps-4', 'ps-99', 'ps-big', 'ps-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12279,7 +12279,7 @@ test('pe', async () => {
       ['pe-1', 'pe-4', 'pe-99', 'pe-big', 'pe-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12320,7 +12320,7 @@ test('pr', async () => {
       ['pr-1', 'pr-4', 'pr-99', 'pr-big', 'pr-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12361,7 +12361,7 @@ test('pb', async () => {
       ['pb-1', 'pb-4', 'pb-99', 'pb-big', 'pb-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12402,7 +12402,7 @@ test('pl', async () => {
       ['pl-1', 'pl-4', 'pl-99', 'pl-big', 'pl-[4px]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --spacing-big: 100rem;
     }
@@ -12594,7 +12594,7 @@ test('font', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --font-sans: ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
       --font-weight-bold: 650;
     }
@@ -12811,7 +12811,7 @@ test('placeholder', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -12936,7 +12936,7 @@ test('decoration', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -13131,7 +13131,7 @@ test('animate', async () => {
       ['animate-spin', 'animate-none', 'animate-[bounce_1s_infinite]', 'animate-not-found'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --animate-spin: spin 1s linear infinite;
     }
 
@@ -13207,7 +13207,7 @@ test('filter', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --blur-xl: 24px;
       --drop-shadow: 0 1px 1px #0000000d;
       --drop-shadow-xl: 0 9px 7px #0000001a;
@@ -13522,7 +13522,7 @@ test('backdrop-filter', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --blur-xl: 24px;
     }
 
@@ -13865,7 +13865,7 @@ test('transition', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --default-transition-timing-function: ease;
       --default-transition-duration: .1s;
       --transition-property: color, background-color, border-color, text-decoration-color, fill, stroke, opacity, box-shadow, transform, filter, backdrop-filter;
@@ -13934,7 +13934,7 @@ test('transition', async () => {
       ['transition', 'transition-all', 'transition-colors'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --default-transition-timing-function: ease;
       --default-transition-duration: .1s;
     }
@@ -14082,7 +14082,7 @@ test('ease', async () => {
       ['ease-in', 'ease-out', 'ease-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --ease-in: cubic-bezier(.4, 0, 1, 1);
       --ease-out: cubic-bezier(0, 0, .2, 1);
     }
@@ -14311,7 +14311,7 @@ test('leading', async () => {
       ['leading-tight', 'leading-6', 'leading-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --leading-tight: 1.25;
       --leading-6: 1.5rem;
     }
@@ -14362,7 +14362,7 @@ test('tracking', async () => {
       ['tracking-normal', 'tracking-wide', 'tracking-[var(--value)]', '-tracking-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --tracking-normal: 0em;
       --tracking-wide: .025em;
     }
@@ -14591,7 +14591,7 @@ test('outline', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -14985,7 +14985,7 @@ test('text', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --spacing: .25rem;
       --color-red-500: #ef4444;
       --text-sm: .875rem;
@@ -15223,7 +15223,7 @@ test('shadow', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
       --shadow-sm: 0 1px 3px 0 #0000001a, 0 1px 2px -1px #0000001a;
       --shadow-xl: 0 20px 25px -5px #0000001a, 0 8px 10px -6px #0000001a;
@@ -15447,7 +15447,7 @@ test('inset-shadow', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
       --inset-shadow: inset 0 2px 4px #0000000d;
       --inset-shadow-sm: inset 0 1px 1px #0000000d;
@@ -15674,7 +15674,7 @@ test('ring', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -15930,7 +15930,7 @@ test('inset-ring', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -16179,7 +16179,7 @@ test('ring-offset', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ":root {
+    ":root, :host {
       --color-red-500: #ef4444;
     }
 
@@ -16357,7 +16357,7 @@ describe('spacing utilities', () => {
     let compiled = build(['px-1', 'px-4'])
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --spacing-4: 1rem;
       }
 
@@ -16378,7 +16378,7 @@ describe('spacing utilities', () => {
     let compiled = build(['px-1', 'px-4'])
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --spacing-4: 1rem;
       }
 
@@ -16398,7 +16398,7 @@ describe('spacing utilities', () => {
     let compiled = build(['px-0.25', 'px-1.5', 'px-2.75', 'px-0.375', 'px-2.50', 'px-.75'])
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --spacing: 4px;
       }
 
@@ -16439,7 +16439,7 @@ describe('spacing utilities', () => {
     let compiled = build(['w-sm', 'max-w-sm', 'min-w-sm', 'basis-sm'])
 
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --spacing-sm: 8px;
         --container-sm: 256px;
       }
@@ -17416,7 +17416,7 @@ describe('custom utilities', () => {
       `
 
       expect(await compileCss(input, ['example-12'])).toMatchInlineSnapshot(`
-        ":root {
+        ":root, :host {
           --spacing: 4px;
         }
 
@@ -17664,7 +17664,7 @@ describe('custom utilities', () => {
     `
 
     expect(await compileCss(input, ['tab-github'])).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --tab-size-github: 8;
       }
 
@@ -17708,7 +17708,7 @@ describe('custom utilities', () => {
     `
 
     expect(await compileCss(input, ['tab-github'])).toMatchInlineSnapshot(`
-      ":root {
+      ":root, :host {
         --tab-size-github: 8;
       }